refacotr: ai chat page folder strcture
