+++ frontend/appflowy_flutter/lib/plugins/ai_chat/presentation/chat_page/chat_animation_list_widget.dart
@@ -1,0 +1,65 @@
+import 'package:appflowy/ai/ai.dart';
+import 'package:appflowy/plugins/ai_chat/application/ai_chat_prelude.dart';
+import 'package:appflowy/plugins/ai_chat/presentation/animated_chat_list.dart';
+import 'package:appflowy/plugins/ai_chat/presentation/chat_welcome_page.dart';
+import 'package:appflowy/plugins/ai_chat/presentation/layout_define.dart';
+import 'package:appflowy_backend/protobuf/flowy-user/protobuf.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:flutter_chat_core/flutter_chat_core.dart';
+
+class ChatAnimationListWidget extends StatelessWidget {
+  const ChatAnimationListWidget({
+    super.key,
+    required this.userProfile,
+    required this.scrollController,
+    required this.itemBuilder,
+  });
+
+  final UserProfilePB userProfile;
+  final ScrollController scrollController;
+  final ChatItem itemBuilder;
+
+  @override
+  Widget build(BuildContext context) {
+    final bloc = context.read<ChatBloc>();
+
+    if (bloc.chatController.messages.isEmpty) {
+      return ChatWelcomePage(
+        userProfile: userProfile,
+        onSelectedQuestion: (question) {
+          final aiPromptInputBloc = context.read<AIPromptInputBloc>();
+          final showPredefinedFormats =
+              aiPromptInputBloc.state.showPredefinedFormats;
+          final predefinedFormat = aiPromptInputBloc.state.predefinedFormat;
+          bloc.add(
+            ChatEvent.sendMessage(
+              message: question,
+              format: showPredefinedFormats ? predefinedFormat : null,
+            ),
+          );
+        },
+      );
+    }
+
+    context
+        .read<ChatSelectMessageBloc>()
+        .add(ChatSelectMessageEvent.enableStartSelectingMessages());
+
+    return BlocSelector<ChatSelectMessageBloc, ChatSelectMessageState, bool>(
+      selector: (state) => state.isSelectingMessages,
+      builder: (context, isSelectingMessages) {
+        return ChatAnimatedListReversed(
+          scrollController: scrollController,
+          itemBuilder: itemBuilder,
+          bottomPadding: isSelectingMessages
+              ? 48.0 + DesktopAIChatSizes.messageActionBarIconSize
+              : 8.0,
+          onLoadPreviousMessages: () {
+            bloc.add(const ChatEvent.loadPreviousMessages());
+          },
+        );
+      },
+    );
+  }
+}
